Extract Path.Builder to own file, refine step builder protocol names
